Refactor for v4 of FatFree Framework
