+++ composer.json
@@ -6,7 +6,7 @@
     "require": {
         "php": "^7.2 || ^8.0",
         "psr/event-dispatcher": "^1.0",
-        "bcosca/fatfree-core": "^3.5"
+        "bcosca/fatfree-core": "v4.x-dev"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.0 || ^9.3",
@@ -16,7 +16,7 @@
         "psr/event-dispatcher-implementation": "1.0"
     },
     "autoload": {
-        "classmap": [ "lib/" ]
+        "psr-4": { "F3\\EventDispatcher\\": "lib/"}
     },
     "autoload-dev": {
         "classmap": [ "tests/" ]
